Extract event candidate extraction module
