+++ backend/app/services/match_unification/resolution.py
@@ -5,7 +5,6 @@ from dataclasses import dataclass, field
 from datetime import datetime
 import hashlib
 import logging
-import re
 import time
 
 from rapidfuzz import fuzz
@@ -23,30 +22,32 @@ from ...models.schemas import (
     EventReviewCaseIn,
     NormalizedOdds,
     NormalizedOutcomeOffer,
-    RawOddsData,
-    RawOutcomeOffer,
     ResolvedEventIn,
     ResolvedEventMemberIn,
 )
 from ...store import odds_store
-from ..league_registry import resolve_league
-from ..normalizer import generate_match_id
-from ..outcome_normalizer import (
-    _build_football_event_resolutions,
-    _event_key_from_raw,
-    FootballEventResolutionMap,
+from .candidate_extraction import (
+    _EventCandidateExtractionStats,
+    _raw_odds_sources,
+    extract_event_candidates,
+)
+from .event_matching import (
+    EventCandidate,
+    _OrientationScore,
+    _TARGETED_SPORTS_FOR_AGGRESSIVE_MERGE,
+    _expand_dotted_token,
+    _orientation_scores,
 )
 from .team_text import (
-    AGGRESSIVE_MERGE_SPORTS as _AGGRESSIVE_MERGE_SPORTS,
     comparison_team_text as _comparison_team_text,
     same_team_context as _same_team_context,
     team_qualifiers as _team_qualifiers,
-    team_similarity as _team_similarity,
 )
 from ..tennis_name_matcher import (
     TENNIS_BROAD_DRIFT_MINUTES,
     tennis_competitor_pair_matches,
 )
+from ..outcome_normalizer import _build_football_event_resolutions
 from ..text_normalizer import normalize_identity_text
 
 logger = logging.getLogger(__name__)
@@ -58,7 +59,6 @@ _HIGH_FUZZY_AVG_SCORE_NON_SUBSET = 90.0
 _HIGH_FUZZY_SIDE_SCORE_NON_SUBSET = 82.0
 _REVIEW_FUZZY_AVG_SCORE = 65.0
 _FUZZY_ORIENTATION_MARGIN = 8.0
-_SOURCE_MATCH_MIN_SCORE = 60.0
 CANONICAL_TEAM_AUTO_MERGE_THRESHOLD = 88.0
 # Anchored low-confidence merge: applies only when the two groups are at the
 # exact same (sport, start_time) slot AND a non-fuzzy corroborator is present
@@ -416,84 +416,6 @@ def _normalize_merge_pairings(
 
 
 @dataclass(frozen=True)
-class EventCandidate:
-    match_id: str
-    bookmaker_id: str
-    sport: str
-    start_time: str
-    home_team_id: int | None
-    away_team_id: int | None
-    home_team: str
-    away_team: str
-    source_league_id: str | None = None
-    source_league_name: str | None = None
-    source_home_team: str | None = None
-    source_away_team: str | None = None
-    source_start_time: str | None = None
-    source_url: str | None = None
-    source_kind: str = "normalized"
-
-    @property
-    def bookmaker_member_key(self) -> tuple[str, str]:
-        return (self.match_id, self.bookmaker_id)
-
-    @property
-    def exact_event_key(self) -> tuple[str, str, tuple[int, int] | tuple[str, str]]:
-        if self.home_team_id is not None and self.away_team_id is not None:
-            team_key: tuple[int, int] | tuple[str, str] = tuple(
-                sorted((self.home_team_id, self.away_team_id))
-            )
-        else:
-            team_key = tuple(
-                sorted(
-                    (
-                        normalize_identity_text(self.home_team),
-                        normalize_identity_text(self.away_team),
-                    )
-                )
-            )
-        return (self.sport, self.start_time, team_key)
-
-
-@dataclass(frozen=True)
-class _RawEventSource:
-    bookmaker_id: str
-    sport: str
-    start_time: str
-    home_team: str
-    away_team: str
-    league_id: str
-    league_name: str
-    source_url: str | None
-    source_kind: str
-
-
-@dataclass(frozen=True)
-class _SourceSlotIndex:
-    all_sources: tuple[_RawEventSource, ...]
-    by_source_url: dict[str, tuple[_RawEventSource, ...]]
-    by_listed_pair: dict[tuple[str, str], tuple[_RawEventSource, ...]]
-    by_unordered_pair: dict[frozenset[str], tuple[_RawEventSource, ...]]
-    source_urls: frozenset[str]
-    league_ids: frozenset[str]
-
-
-@dataclass(frozen=True)
-class _OrientationScore:
-    orientation: str
-    home_score: float
-    away_score: float
-
-    @property
-    def avg_score(self) -> float:
-        return (self.home_score + self.away_score) / 2
-
-    @property
-    def weak_side_score(self) -> float:
-        return min(self.home_score, self.away_score)
-
-
-@dataclass(frozen=True)
 class _PairResolution:
     confidence: float
     score: float
@@ -537,46 +459,6 @@ class EventResolutionGroup:
     confidence: float
     members: tuple[EventCandidate, ...]
     evidence: tuple[str, ...]
-
-
-@dataclass
-class _EventCandidateExtractionStats:
-    extract_raw_odds_sources_ms: int = 0
-    extract_raw_outcome_sources_ms: int = 0
-    extract_normalized_odds_candidates_ms: int = 0
-    extract_normalized_outcome_candidates_ms: int = 0
-    raw_odds_rows_scanned: int = 0
-    raw_odds_sources_emitted: int = 0
-    raw_outcome_offer_rows_scanned: int = 0
-    raw_outcome_sources_emitted: int = 0
-    normalized_odds_rows_scanned: int = 0
-    normalized_odds_candidates_emitted: int = 0
-    normalized_outcome_offer_rows_scanned: int = 0
-    normalized_outcome_candidates_emitted: int = 0
-    stored_outcome_match_bookmaker_count: int = 0
-    source_match_lookup_count: int = 0
-    source_match_source_count: int = 0
-    source_match_scored_source_count: int = 0
-    source_match_index_candidate_count: int = 0
-    source_match_exact_url_hit_count: int = 0
-    source_match_listed_pair_hit_count: int = 0
-    source_match_unordered_pair_hit_count: int = 0
-    source_match_fallback_scan_count: int = 0
-    source_match_max_sources_per_lookup: int = 0
-    source_match_slot_lookup_counts: Counter[tuple[str, str, str]] = field(
-        default_factory=Counter
-    )
-    source_match_slot_source_counts: Counter[tuple[str, str, str]] = field(
-        default_factory=Counter
-    )
-    football_raw_candidate_count: int = 0
-    football_raw_resolution_candidates_ms: int = 0
-    reused_football_event_resolution_count: int = 0
-    _source_match_seconds: float = 0.0
-
-    @property
-    def source_match_ms(self) -> int:
-        return int(self._source_match_seconds * 1000)
 
 
 @dataclass
@@ -1373,264 +1255,22 @@ def _split_sport_diagnostics(
     )
 
 
-def _league_source(raw_league_id: str, bookmaker_id: str) -> tuple[str, str]:
-    resolution = resolve_league(raw_league_id, bookmaker_id=bookmaker_id)
-    return resolution.league_id, resolution.display_name
-
-
-def _league_source_cached(
-    raw_league_id: str,
-    bookmaker_id: str,
-    cache: dict[tuple[str, str], tuple[str, str]] | None,
-) -> tuple[str, str]:
-    if cache is None:
-        return _league_source(raw_league_id, bookmaker_id)
-    key = (raw_league_id, bookmaker_id)
-    value = cache.get(key)
-    if value is None:
-        value = _league_source(raw_league_id, bookmaker_id)
-        cache[key] = value
-    return value
-
-
-def _normalized_identity_cached(value: str, cache: dict[str, str]) -> str:
-    normalized = cache.get(value)
-    if normalized is None:
-        normalized = normalize_identity_text(value)
-        cache[value] = normalized
-    return normalized
-
-
-def _raw_odds_sources(
-    raw_odds: list[RawOddsData],
-    *,
-    league_cache: dict[tuple[str, str], tuple[str, str]] | None = None,
-) -> list[_RawEventSource]:
-    source_rows: dict[tuple[str, str, str, str, str, str | None], RawOddsData] = {}
-    identity_cache: dict[str, str] = {}
-    for raw in raw_odds:
-        if not raw.start_time:
-            continue
-        key = (
-            raw.bookmaker_id,
-            raw.sport,
-            raw.start_time,
-            _normalized_identity_cached(raw.home_team, identity_cache),
-            _normalized_identity_cached(raw.away_team, identity_cache),
-            raw.source_url,
-        )
-        source_rows.setdefault(key, raw)
-
-    sources: list[_RawEventSource] = []
-    for raw in source_rows.values():
-        league_id, league_name = _league_source_cached(
-            raw.league_id,
-            raw.bookmaker_id,
-            league_cache,
-        )
-        sources.append(
-            _RawEventSource(
-                bookmaker_id=raw.bookmaker_id,
-                sport=raw.sport,
-                start_time=raw.start_time,
-                home_team=raw.home_team,
-                away_team=raw.away_team,
-                league_id=league_id,
-                league_name=league_name,
-                source_url=raw.source_url,
-                source_kind="raw_odds",
-            ),
-        )
-    return sources
-
-
-def _raw_outcome_sources(
-    raw_offers: list[RawOutcomeOffer],
-    *,
-    league_cache: dict[tuple[str, str], tuple[str, str]] | None = None,
-) -> list[_RawEventSource]:
-    source_rows: dict[tuple[str, str, str, str, str, str | None], RawOutcomeOffer] = {}
-    identity_cache: dict[str, str] = {}
-    for raw in raw_offers:
-        if not raw.start_time:
-            continue
-        key = (
-            raw.bookmaker_id,
-            raw.sport,
-            raw.start_time,
-            _normalized_identity_cached(raw.home_team, identity_cache),
-            _normalized_identity_cached(raw.away_team, identity_cache),
-            raw.source_url,
-        )
-        source_rows.setdefault(key, raw)
-
-    sources: list[_RawEventSource] = []
-    for raw in source_rows.values():
-        league_id, league_name = _league_source_cached(
-            raw.league_id,
-            raw.bookmaker_id,
-            league_cache,
-        )
-        sources.append(
-            _RawEventSource(
-                bookmaker_id=raw.bookmaker_id,
-                sport=raw.sport,
-                start_time=raw.start_time,
-                home_team=raw.home_team,
-                away_team=raw.away_team,
-                league_id=league_id,
-                league_name=league_name,
-                source_url=raw.source_url,
-                source_kind="raw_outcome_offer",
-            ),
-        )
-    return sources
-
-
-# Sports for which the Match Unification resolver activates aggressive aliasing & dot-expansion
-# heuristics. Re-exported alias of the Match Unification team-text rules so the
-# modules cannot drift; new sports must be enabled in exactly one place.
-_TARGETED_SPORTS_FOR_AGGRESSIVE_MERGE: frozenset[str] = _AGGRESSIVE_MERGE_SPORTS
-
-# 2-letter dot-prefixes that overlap heavily with non-team words (street,
-# fort, mount, port, point, doctor, mister, avenue, saint) and would
-# otherwise produce false-positive expansions even within basketball
-# (e.g. ``St.Petersburg`` ↔ ``Stockholm Petersburg``). Keeping these out of
-# the dot-expansion logic preserves the genuine ``Ch.More`` ↔ ``Cherno More``
-# case while blocking the geographic collision class.
-_AMBIGUOUS_DOT_PREFIXES: frozenset[str] = frozenset(
-    {"st", "ft", "mt", "pt", "dr", "mr", "av"}
-)
-
-
-def _expand_dotted_token(name: str, counterpart: str) -> str:
-    """Substitute dot-truncated tokens (``Ch.``, ``Pl.``, ``Ch.More``) by an
-    unambiguous expansion drawn from ``counterpart``.
-
-    Only used inside the Match Unification — keeps shared :func:`_team_similarity`
-    untouched so football pairing is unaffected. Restrictions:
-
-    * Token must end with ``.`` and have at least 2 characters of prefix
-      (1-letter prefixes are too ambiguous, e.g. ``B.`` could be Bayern,
-      Brest, Belgrade, …).
-    * The prefix must not be in ``_AMBIGUOUS_DOT_PREFIXES`` — these short
-      geographic / honorific prefixes (``St``, ``Mt``, ``Ft``, …) collide
-      with real team-name tokens (``Stockholm``, ``Manchester``, ``Fort``,
-      …) and the structural anchor check below cannot disambiguate them.
-    * The counterpart must contain exactly one token starting with that
-      prefix; ambiguous expansions are dropped.
-    * The source name must contain at least one OTHER non-dotted token that
-      already appears in the counterpart — this anchors the expansion in
-      genuine name overlap and blocks coincidences like
-      ``St. Petersburg`` ↔ ``Stockholm Giants`` where ``St`` would
-      otherwise expand to ``Stockholm`` purely on prefix uniqueness.
-
-    Compound tokens with internal dots (``Ch.More`` → ``Ch. More``) are
-    pre-split before expansion so a missing space after the period does not
-    mask the abbreviation.
-    """
-
-    spaced = re.sub(r"\.(?=\S)", ". ", name)
-    counterpart_spaced = re.sub(r"\.(?=\S)", ". ", counterpart)
-    counterpart_tokens = counterpart_spaced.split()
-    counterpart_token_set = {token.lower().rstrip(".") for token in counterpart_tokens}
-    source_tokens = spaced.split()
-    has_anchor = any(
-        not token.endswith(".") and token.lower() in counterpart_token_set
-        for token in source_tokens
-    )
-    if not has_anchor:
-        return name
-    output: list[str] = []
-    for token in source_tokens:
-        if not token.endswith(".") or len(token) < 3:
-            output.append(token)
-            continue
-        prefix = token[:-1].lower()
-        if len(prefix) < 2:
-            output.append(token)
-            continue
-        if prefix in _AMBIGUOUS_DOT_PREFIXES:
-            output.append(token)
-            continue
-        candidates = [
-            candidate
-            for candidate in counterpart_tokens
-            if len(candidate) > len(prefix)
-            and candidate.lower().startswith(prefix)
-        ]
-        if len(candidates) == 1:
-            output.append(candidates[0])
-        else:
-            output.append(token)
-    return " ".join(output)
-
-
-def _resolver_team_similarity(
-    left: str, right: str, *, sport: str | None = None
-) -> float:
-    """Match-Unification-local team similarity that pre-expands dot-truncations.
-
-    Equivalent to :func:`_team_similarity` for cases without dotted
-    abbreviations.
-
-    Sport-gated: dot expansion only fires for sports in
-    ``_TARGETED_SPORTS_FOR_AGGRESSIVE_MERGE``. Football has its own
-    pairing flow with stricter handling, and the dot-expansion logic
-    cannot structurally distinguish ``Ch.More`` (basketball — should expand)
-    from ``St.Petersburg`` ↔ ``Stockholm Petersburg`` (football — would
-    incorrectly merge two distinct cities).
-    """
-
-    expanded_left = left
-    expanded_right = right
-    if sport in _TARGETED_SPORTS_FOR_AGGRESSIVE_MERGE:
-        expanded_left = _expand_dotted_token(left, right)
-        expanded_right = _expand_dotted_token(right, left)
-    return _team_similarity(expanded_left, expanded_right, sport=sport)
-
-
-def _orientation_scores(
-    left_home: str,
-    left_away: str,
-    right_home: str,
-    right_away: str,
+def _shared_significant_tokens(
+    left_name: str,
+    right_name: str,
     *,
     sport: str | None = None,
-) -> list[_OrientationScore]:
-    if sport == "tennis":
-        return [
-            _OrientationScore(
-                orientation=match.orientation,
-                home_score=match.home_score,
-                away_score=match.away_score,
-            )
-            for match in tennis_competitor_pair_matches(
-                left_home,
-                left_away,
-                right_home,
-                right_away,
-            )
-        ]
+    text_cache: _ResolverTextCache | None = None,
+) -> set[str]:
+    if text_cache is not None:
+        return text_cache.significant_tokens(
+            left_name, sport=sport
+        ) & text_cache.significant_tokens(right_name, sport=sport)
+    return _significant_team_tokens(left_name, sport=sport) & _significant_team_tokens(
+        right_name,
+        sport=sport,
+    )
 
-    scores: list[_OrientationScore] = []
-    if _same_team_context(left_home, right_home, sport=sport) and _same_team_context(left_away, right_away, sport=sport):
-        scores.append(
-            _OrientationScore(
-                orientation="as_listed",
-                home_score=_resolver_team_similarity(left_home, right_home, sport=sport),
-                away_score=_resolver_team_similarity(left_away, right_away, sport=sport),
-            )
-        )
-    if _same_team_context(left_home, right_away, sport=sport) and _same_team_context(left_away, right_home, sport=sport):
-        scores.append(
-            _OrientationScore(
-                orientation="reversed",
-                home_score=_resolver_team_similarity(left_home, right_away, sport=sport),
-                away_score=_resolver_team_similarity(left_away, right_home, sport=sport),
-            )
-        )
-    return sorted(scores, key=lambda score: score.avg_score, reverse=True)
 
 def _is_subset_or_equal_token_pair(
     left_name: str,
@@ -1696,534 +1336,6 @@ def _matching_canonical_sides(
     if away_pair[0] is not None and away_pair[0] == away_pair[1]:
         matching.add("away")
     return matching
-
-
-def _source_match_score(source: _RawEventSource, candidate: EventCandidate) -> float:
-    scores = _orientation_scores(
-        source.home_team,
-        source.away_team,
-        candidate.home_team,
-        candidate.away_team,
-        sport=source.sport,
-    )
-    if not scores:
-        return 0.0
-    score = scores[0].avg_score
-    if source.source_url and source.source_url == candidate.source_url:
-        score += 10.0
-    if source.league_id == candidate.source_league_id:
-        score += 3.0
-    return score
-
-
-def _source_match_max_score(
-    candidate: EventCandidate,
-    slot_index: _SourceSlotIndex,
-) -> float:
-    score = 100.0
-    if candidate.source_url and candidate.source_url in slot_index.source_urls:
-        score += 10.0
-    if (
-        candidate.source_league_id
-        and candidate.source_league_id in slot_index.league_ids
-    ):
-        score += 3.0
-    return score
-
-
-def _source_listed_pair_key(
-    home_team: str | None,
-    away_team: str | None,
-) -> tuple[str, str]:
-    return (
-        normalize_identity_text(home_team),
-        normalize_identity_text(away_team),
-    )
-
-
-def _source_unordered_pair_key(
-    home_team: str | None,
-    away_team: str | None,
-) -> frozenset[str]:
-    return frozenset(_source_listed_pair_key(home_team, away_team))
-
-
-def _freeze_multimap(
-    rows: dict,
-) -> dict:
-    return {key: tuple(value) for key, value in rows.items()}
-
-
-def _build_source_slot_indexes(
-    sources: list[_RawEventSource],
-) -> dict[tuple[str, str, str], _SourceSlotIndex]:
-    by_slot: dict[tuple[str, str, str], list[_RawEventSource]] = defaultdict(list)
-    for source in sources:
-        by_slot[(source.bookmaker_id, source.sport, source.start_time)].append(source)
-
-    indexes: dict[tuple[str, str, str], _SourceSlotIndex] = {}
-    for slot_key, slot_sources in by_slot.items():
-        by_source_url: dict[str, list[_RawEventSource]] = defaultdict(list)
-        by_listed_pair: dict[tuple[str, str], list[_RawEventSource]] = defaultdict(list)
-        by_unordered_pair: dict[frozenset[str], list[_RawEventSource]] = defaultdict(list)
-        source_urls: set[str] = set()
-        league_ids: set[str] = set()
-        for source in slot_sources:
-            if source.source_url:
-                by_source_url[source.source_url].append(source)
-                source_urls.add(source.source_url)
-            league_ids.add(source.league_id)
-            listed_key = _source_listed_pair_key(source.home_team, source.away_team)
-            by_listed_pair[listed_key].append(source)
-            by_unordered_pair[frozenset(listed_key)].append(source)
-        indexes[slot_key] = _SourceSlotIndex(
-            all_sources=tuple(slot_sources),
-            by_source_url=_freeze_multimap(by_source_url),
-            by_listed_pair=_freeze_multimap(by_listed_pair),
-            by_unordered_pair=_freeze_multimap(by_unordered_pair),
-            source_urls=frozenset(source_urls),
-            league_ids=frozenset(league_ids),
-        )
-    return indexes
-
-
-def _best_source(
-    sources_by_slot: dict[tuple[str, str, str], _SourceSlotIndex],
-    candidate: EventCandidate,
-    *,
-    stats: _EventCandidateExtractionStats | None = None,
-) -> _RawEventSource | None:
-    started_at = time.perf_counter()
-    slot_key = (candidate.bookmaker_id, candidate.sport, candidate.start_time)
-    slot_index = sources_by_slot.get(slot_key)
-    sources = slot_index.all_sources if slot_index is not None else ()
-    if stats is not None:
-        stats.source_match_lookup_count += 1
-        stats.source_match_source_count += len(sources)
-        stats.source_match_max_sources_per_lookup = max(
-            stats.source_match_max_sources_per_lookup,
-            len(sources),
-        )
-        stats.source_match_slot_lookup_counts[slot_key] += 1
-        stats.source_match_slot_source_counts[slot_key] += len(sources)
-    try:
-        if slot_index is None:
-            return None
-        score_cache: dict[int, float] = {}
-
-        def score_source(source: _RawEventSource) -> float:
-            cache_key = id(source)
-            if cache_key not in score_cache:
-                score_cache[cache_key] = _source_match_score(source, candidate)
-                if stats is not None:
-                    stats.source_match_scored_source_count += 1
-            return score_cache[cache_key]
-
-        def best_from_subset(
-            subset: tuple[_RawEventSource, ...],
-        ) -> tuple[_RawEventSource | None, float]:
-            best_source: _RawEventSource | None = None
-            best_subset_score = 0.0
-            for source in subset:
-                score = score_source(source)
-                if best_source is None or score > best_subset_score:
-                    best_source = source
-                    best_subset_score = score
-            return best_source, best_subset_score
-
-        def maybe_fast_return(
-            subset: tuple[_RawEventSource, ...],
-            counter_name: str,
-        ) -> _RawEventSource | None:
-            if not subset:
-                return None
-            if stats is not None:
-                stats.source_match_index_candidate_count += len(subset)
-            best_subset_source, best_subset_score = best_from_subset(subset)
-            if best_subset_source is None:
-                return None
-            if best_subset_score < _SOURCE_MATCH_MIN_SCORE:
-                return None
-            if best_subset_score < _source_match_max_score(candidate, slot_index):
-                return None
-            if not (
-                candidate.source_url
-                and best_subset_source.source_url == candidate.source_url
-            ) and (not sources or sources[0] is not best_subset_source):
-                return None
-            if stats is not None:
-                setattr(stats, counter_name, getattr(stats, counter_name) + 1)
-            return best_subset_source
-
-        if candidate.source_url:
-            source = maybe_fast_return(
-                slot_index.by_source_url.get(candidate.source_url, ()),
-                "source_match_exact_url_hit_count",
-            )
-            if source is not None:
-                return source
-
-        listed_key = _source_listed_pair_key(candidate.home_team, candidate.away_team)
-        source = maybe_fast_return(
-            slot_index.by_listed_pair.get(listed_key, ()),
-            "source_match_listed_pair_hit_count",
-        )
-        if source is not None:
-            return source
-
-        unordered_key = _source_unordered_pair_key(
-            candidate.home_team,
-            candidate.away_team,
-        )
-        source = maybe_fast_return(
-            slot_index.by_unordered_pair.get(unordered_key, ()),
-            "source_match_unordered_pair_hit_count",
-        )
-        if source is not None:
-            return source
-
-        if stats is not None:
-            stats.source_match_fallback_scan_count += 1
-        best: _RawEventSource | None = None
-        best_score = 0.0
-        for source in sources:
-            score = score_source(source)
-            if best is None or score > best_score:
-                best = source
-                best_score = score
-        if best is None or best_score < _SOURCE_MATCH_MIN_SCORE:
-            return None
-        return best
-    finally:
-        if stats is not None:
-            stats._source_match_seconds += time.perf_counter() - started_at
-
-
-def _normalized_odds_candidate(
-    row: NormalizedOdds,
-    source: _RawEventSource | None,
-    *,
-    league_cache: dict[tuple[str, str], tuple[str, str]] | None = None,
-) -> EventCandidate | None:
-    if not row.start_time:
-        return None
-    league_id = row.league_id
-    league_name = _league_source_cached(
-        row.league_id,
-        row.bookmaker_id,
-        league_cache,
-    )[1]
-    if source is not None:
-        league_id = source.league_id
-        league_name = source.league_name
-    return EventCandidate(
-        match_id=row.match_id,
-        bookmaker_id=row.bookmaker_id,
-        sport=row.sport,
-        start_time=row.start_time,
-        home_team_id=row.home_team_id or None,
-        away_team_id=row.away_team_id or None,
-        home_team=row.home_team,
-        away_team=row.away_team,
-        source_league_id=league_id,
-        source_league_name=league_name,
-        source_home_team=source.home_team if source else row.home_team,
-        source_away_team=source.away_team if source else row.away_team,
-        source_start_time=source.start_time if source else row.start_time,
-        source_url=source.source_url if source and source.source_url else row.source_url,
-        source_kind=source.source_kind if source else "normalized_odds",
-    )
-
-
-def _normalized_outcome_candidate(
-    row: NormalizedOutcomeOffer,
-    source: _RawEventSource | None,
-    *,
-    league_cache: dict[tuple[str, str], tuple[str, str]] | None = None,
-) -> EventCandidate | None:
-    if not row.start_time:
-        return None
-    league_id = row.league_id
-    league_name = _league_source_cached(
-        row.league_id,
-        row.bookmaker_id,
-        league_cache,
-    )[1]
-    if source is not None:
-        league_id = source.league_id
-        league_name = source.league_name
-    return EventCandidate(
-        match_id=row.match_id,
-        bookmaker_id=row.bookmaker_id,
-        sport=row.sport,
-        start_time=row.start_time,
-        home_team_id=row.home_team_id or None,
-        away_team_id=row.away_team_id or None,
-        home_team=row.home_team,
-        away_team=row.away_team,
-        source_league_id=league_id,
-        source_league_name=league_name,
-        source_home_team=source.home_team if source else row.home_team,
-        source_away_team=source.away_team if source else row.away_team,
-        source_start_time=source.start_time if source else row.start_time,
-        source_url=source.source_url if source and source.source_url else row.source_url,
-        source_kind=source.source_kind if source else "normalized_outcome_offer",
-    )
-
-
-def _merge_candidate(
-    candidates: dict[tuple[str, str], EventCandidate],
-    candidate: EventCandidate | None,
-) -> None:
-    if candidate is None:
-        return
-    existing = candidates.get(candidate.bookmaker_member_key)
-    if existing is None:
-        candidates[candidate.bookmaker_member_key] = candidate
-        return
-    if existing.source_kind.startswith(
-        "normalized"
-    ) and not candidate.source_kind.startswith("normalized"):
-        candidates[candidate.bookmaker_member_key] = candidate
-        return
-    if existing.source_url is None and candidate.source_url is not None:
-        candidates[candidate.bookmaker_member_key] = candidate
-
-
-def _prefer_row_with_source_url(existing, row):
-    if not getattr(existing, "source_url", None) and getattr(
-        row,
-        "source_url",
-        None,
-    ):
-        return row
-    return existing
-
-
-def _unique_normalized_odds_rows(
-    rows: list[NormalizedOdds],
-) -> list[NormalizedOdds]:
-    unique: dict[tuple[str, str], NormalizedOdds] = {}
-    for row in rows:
-        if not row.start_time:
-            continue
-        key = (row.match_id, row.bookmaker_id)
-        existing = unique.get(key)
-        unique[key] = (
-            row if existing is None else _prefer_row_with_source_url(existing, row)
-        )
-    return list(unique.values())
-
-
-def _unique_normalized_outcome_rows(
-    rows: list[NormalizedOutcomeOffer],
-) -> list[NormalizedOutcomeOffer]:
-    unique: dict[tuple[str, str], NormalizedOutcomeOffer] = {}
-    for row in rows:
-        if not row.start_time:
-            continue
-        key = (row.match_id, row.bookmaker_id)
-        existing = unique.get(key)
-        unique[key] = (
-            row if existing is None else _prefer_row_with_source_url(existing, row)
-        )
-    return list(unique.values())
-
-
-def _football_raw_resolution_candidates(
-    raw_offers: list[RawOutcomeOffer],
-    stored_match_bookmakers: set[tuple[str, str]],
-    *,
-    football_event_resolutions: FootballEventResolutionMap | None = None,
-) -> list[EventCandidate]:
-    if not raw_offers:
-        return []
-    event_resolutions = (
-        football_event_resolutions
-        if football_event_resolutions is not None
-        else _build_football_event_resolutions(raw_offers)
-    )
-    seen_raw_events: set[tuple[str, str, str, str, str]] = set()
-    candidates: list[EventCandidate] = []
-    for raw in raw_offers:
-        event_key = _event_key_from_raw(raw)
-        if event_key is None or event_key in seen_raw_events:
-            continue
-        seen_raw_events.add(event_key)
-        resolution = event_resolutions.get(event_key)
-        if resolution is None:
-            continue
-        match_id = generate_match_id(
-            resolution.slot.home_team_id,
-            resolution.slot.away_team_id,
-            resolution.slot.start_time,
-            raw.sport,
-        )
-        if (match_id, raw.bookmaker_id) not in stored_match_bookmakers:
-            continue
-        league_id, league_name = _league_source(raw.league_id, raw.bookmaker_id)
-        candidates.append(
-            EventCandidate(
-                match_id=match_id,
-                bookmaker_id=raw.bookmaker_id,
-                sport=raw.sport,
-                start_time=resolution.slot.start_time,
-                home_team_id=resolution.slot.home_team_id,
-                away_team_id=resolution.slot.away_team_id,
-                home_team=resolution.slot.home_team,
-                away_team=resolution.slot.away_team,
-                source_league_id=league_id,
-                source_league_name=league_name,
-                source_home_team=raw.home_team,
-                source_away_team=raw.away_team,
-                source_start_time=raw.start_time,
-                source_url=raw.source_url,
-                source_kind="raw_outcome_offer",
-            )
-        )
-    return candidates
-
-
-def extract_event_candidates(
-    *,
-    raw_odds: list[RawOddsData],
-    raw_outcome_offers: list[RawOutcomeOffer],
-    normalized_odds: list[NormalizedOdds],
-    normalized_outcome_offers: list[NormalizedOutcomeOffer],
-    football_event_resolutions: FootballEventResolutionMap | None = None,
-    stats: _EventCandidateExtractionStats | None = None,
-) -> list[EventCandidate]:
-    """Build one source-event candidate per bookmaker/match from the current scrape."""
-
-    candidates: dict[tuple[str, str], EventCandidate] = {}
-    league_cache: dict[tuple[str, str], tuple[str, str]] = {}
-
-    raw_odds_sources_started_at = time.perf_counter()
-    odds_sources = _raw_odds_sources(raw_odds, league_cache=league_cache)
-    if stats is not None:
-        stats.extract_raw_odds_sources_ms = _elapsed_ms(raw_odds_sources_started_at)
-        stats.raw_odds_rows_scanned = len(raw_odds)
-        stats.raw_odds_sources_emitted = len(odds_sources)
-    odds_sources_by_slot = _build_source_slot_indexes(odds_sources)
-
-    raw_outcome_sources_started_at = time.perf_counter()
-    outcome_sources = _raw_outcome_sources(
-        raw_outcome_offers,
-        league_cache=league_cache,
-    )
-    if stats is not None:
-        stats.extract_raw_outcome_sources_ms = _elapsed_ms(
-            raw_outcome_sources_started_at
-        )
-        stats.raw_outcome_offer_rows_scanned = len(raw_outcome_offers)
-        stats.raw_outcome_sources_emitted = len(outcome_sources)
-    outcome_sources_by_slot = _build_source_slot_indexes(outcome_sources)
-
-    normalized_odds_started_at = time.perf_counter()
-    unique_normalized_odds = _unique_normalized_odds_rows(normalized_odds)
-    if stats is not None:
-        stats.normalized_odds_rows_scanned = len(normalized_odds)
-        stats.normalized_odds_candidates_emitted = len(unique_normalized_odds)
-    for row in unique_normalized_odds:
-        provisional = EventCandidate(
-            match_id=row.match_id,
-            bookmaker_id=row.bookmaker_id,
-            sport=row.sport,
-            start_time=row.start_time or "",
-            home_team_id=row.home_team_id or None,
-            away_team_id=row.away_team_id or None,
-            home_team=row.home_team,
-            away_team=row.away_team,
-            source_league_id=row.league_id,
-            source_url=row.source_url,
-        )
-        source = _best_source(odds_sources_by_slot, provisional, stats=stats)
-        _merge_candidate(
-            candidates,
-            _normalized_odds_candidate(row, source, league_cache=league_cache),
-        )
-    if stats is not None:
-        stats.extract_normalized_odds_candidates_ms = _elapsed_ms(
-            normalized_odds_started_at
-        )
-
-    normalized_outcome_started_at = time.perf_counter()
-    unique_normalized_outcomes = _unique_normalized_outcome_rows(
-        normalized_outcome_offers
-    )
-    if stats is not None:
-        stats.normalized_outcome_offer_rows_scanned = len(normalized_outcome_offers)
-        stats.normalized_outcome_candidates_emitted = len(unique_normalized_outcomes)
-    for row in unique_normalized_outcomes:
-        provisional = EventCandidate(
-            match_id=row.match_id,
-            bookmaker_id=row.bookmaker_id,
-            sport=row.sport,
-            start_time=row.start_time or "",
-            home_team_id=row.home_team_id or None,
-            away_team_id=row.away_team_id or None,
-            home_team=row.home_team,
-            away_team=row.away_team,
-            source_league_id=row.league_id,
-            source_url=row.source_url,
-        )
-        source = _best_source(outcome_sources_by_slot, provisional, stats=stats)
-        _merge_candidate(
-            candidates,
-            _normalized_outcome_candidate(row, source, league_cache=league_cache),
-        )
-    if stats is not None:
-        stats.extract_normalized_outcome_candidates_ms = _elapsed_ms(
-            normalized_outcome_started_at
-        )
-
-    stored_outcome_match_bookmakers = {
-        (offer.match_id, offer.bookmaker_id) for offer in normalized_outcome_offers
-    }
-    if stats is not None:
-        stats.stored_outcome_match_bookmaker_count = len(stored_outcome_match_bookmakers)
-    football_candidates_started_at = time.perf_counter()
-    football_candidates = _football_raw_resolution_candidates(
-        raw_outcome_offers,
-        stored_outcome_match_bookmakers,
-        football_event_resolutions=football_event_resolutions,
-    )
-    if stats is not None and football_event_resolutions is not None:
-        stats.reused_football_event_resolution_count = len(football_candidates)
-    if stats is not None:
-        stats.football_raw_resolution_candidates_ms = _elapsed_ms(
-            football_candidates_started_at
-        )
-        stats.football_raw_candidate_count = len(football_candidates)
-    for candidate in football_candidates:
-        _merge_candidate(candidates, candidate)
-
-    return sorted(
-        candidates.values(),
-        key=lambda candidate: (
-            candidate.sport,
-            candidate.start_time,
-            candidate.match_id,
-            candidate.bookmaker_id,
-        ),
-    )
-
-
-def _shared_significant_tokens(
-    left_name: str,
-    right_name: str,
-    *,
-    sport: str | None = None,
-    text_cache: _ResolverTextCache | None = None,
-) -> set[str]:
-    if text_cache is not None:
-        return text_cache.significant_tokens(
-            left_name, sport=sport
-        ) & text_cache.significant_tokens(right_name, sport=sport)
-    return _significant_team_tokens(left_name, sport=sport) & _significant_team_tokens(
-        right_name,
-        sport=sport,
-    )
 
 
 def _anchored_low_conf_detail(
